nacl.i: fix py2.6, accept both 'int' and 'long' when taking a number
Without this, nacl.randombytes(32) on py2 fails with a SystemError("bad
argument to internal function"), as it tries to call
PyLong_AsUnsignedLongLongMask() on a (short) int. Instead, we use
PyInt_AsUnsignedLongLongMask(), which accepts both ints and longs.

On py3, there is no PyInt_* API, since everything is a long. We prefer
PyLong_AsUnsignedLongLong (without the -Mask suffix), which throws an
exception on overflow rather than silently truncating.